Test: mapping test

+++ src/main/java/com/fundfun/fundfund/domain/product/Product.java
@@ -2,6 +2,7 @@ package com.fundfun.fundfund.domain.product;
 
 import com.fundfun.fundfund.base.BaseTimeEntity;
 import com.fundfun.fundfund.domain.order.Orders;
+import com.fundfun.fundfund.domain.user.Users;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
@@ -27,14 +28,19 @@ public class Product extends BaseTimeEntity {
     private String title;
     private String crowdStart;
     private String crowdEnd;
-    private Integer goal;
-    private Integer currentGoal;
+    private Long goal;
+    private Long currentGoal;
     private String status;
     private String description;
+
+//    @OneToOne
+//    @JoinColumn(name = "order_id")
+//    private Orders orders;
 
 //    @OneToMany(mappedBy = "orders")
 //    @JoinColumn(name = "order_id")
 //    private List<Orders> orders;
+
 
 //    @ManyToOne
 //    @JoinColumn(name = "user_id")
@@ -43,11 +49,7 @@ public class Product extends BaseTimeEntity {
         //UUID encode
         Base64.Encoder encoder = Base64.getEncoder();
         String encodedString = encoder.encodeToString(this.id.toString().getBytes());
-//        ByteBuffer bb = ByteBuffer.wrap(new byte[16]);
-//        bb.putLong(this.id.getMostSignificantBits());
-//        bb.putLong(this.id.getLeastSignificantBits());
-//        byte[] encodedBits = encoder.encode(bb.array());
-//        System.out.println("encodedBits: " + encodedBits);
+
         return encodedString;
     }
